backend/requirements.txt
flask
flask-cors
flask-limiter
python-dotenv
orjson
gunicorn
openai
